requirements.txt
dash
dash_core_components
dash_html_components
plotly.express

pandas
